enhancement: melhorias  #50
