+++ src/main/java/com/softawii/capivara/core/DroneManager.java
@@ -1,27 +1,40 @@
 package com.softawii.capivara.core;
 
 import com.softawii.capivara.entity.VoiceDrone;
+import com.softawii.capivara.entity.VoiceHive;
 import com.softawii.capivara.exceptions.InvalidInputException;
 import com.softawii.capivara.exceptions.KeyNotFoundException;
+import com.softawii.capivara.listeners.VoiceGroup;
 import com.softawii.capivara.services.VoiceDroneService;
+import com.softawii.capivara.services.VoiceHiveService;
 import com.softawii.curupira.exceptions.MissingPermissionsException;
+import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.*;
 import net.dv8tion.jda.api.entities.channel.unions.MessageChannelUnion;
 import net.dv8tion.jda.api.events.interaction.ModalInteractionEvent;
+import net.dv8tion.jda.api.interactions.components.ActionRow;
 import net.dv8tion.jda.api.interactions.components.Modal;
+import net.dv8tion.jda.api.interactions.components.buttons.Button;
 import net.dv8tion.jda.api.interactions.components.text.TextInput;
 import net.dv8tion.jda.api.interactions.components.text.TextInputStyle;
+import net.dv8tion.jda.api.interactions.modals.ModalMapping;
+import net.dv8tion.jda.api.requests.RestAction;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.jetbrains.annotations.Nullable;
 import org.springframework.stereotype.Component;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Objects;
+import java.awt.*;
+import java.util.*;
+import java.util.List;
 
 @Component
 public class DroneManager {
 
+    private final Logger LOGGER = LogManager.getLogger(VoiceManager.class);
     private final VoiceDroneService voiceDroneService;
+    private final VoiceHiveService  voiceHiveService;
     private final VoiceManager voiceManager;
 
     private final String renameDrone     = "drone-manager-rename";
@@ -29,17 +42,11 @@ public class DroneManager {
     private final String connectDrone    = "drone-manager-connect";
     private final String visibilityDrone = "drone-manager-visibility";
 
-    public DroneManager(VoiceDroneService voiceDroneService, VoiceManager voiceManager) {
+
+    public DroneManager(VoiceDroneService voiceDroneService, VoiceHiveService voiceHiveService, VoiceManager voiceManager) {
         this.voiceDroneService = voiceDroneService;
+        this.voiceHiveService = voiceHiveService;
         this.voiceManager = voiceManager;
-    }
-
-    private boolean canConnect(VoiceChannel channel) {
-        return channel.getGuild().getPublicRole().getPermissions(channel).contains(Permission.VOICE_CONNECT);
-    }
-
-    private boolean isVisible(VoiceChannel channel) {
-        return channel.getGuild().getPublicRole().getPermissions(channel).contains(Permission.VIEW_CHANNEL);
     }
 
     public boolean canInteract(VoiceChannel channel, Member member) throws KeyNotFoundException {
@@ -136,11 +143,12 @@ public class DroneManager {
         if (newVisibilityDrone.equals("visible")) give.add(Permission.VIEW_CHANNEL);
         else deny.add(Permission.VIEW_CHANNEL);
 
-        voiceChannel.getManager().setName(newName).setUserLimit(newLimitDrone).queue();
-        voiceChannel.upsertPermissionOverride(publicRole).deny(deny).grant(give).queue();
+        RestAction<Void> action = voiceChannel.getManager().setName(newName).setUserLimit(newLimitDrone)
+                .and(voiceChannel.upsertPermissionOverride(publicRole).deny(deny).grant(give));
         if(textChannel != null) {
-            textChannel.getManager().setName(newName).queue();
+            action = action.and(textChannel.getManager().setName(newName));
         }
+        action.queue();
     }
 
     public void checkToChangeChatAccess(VoiceChannel channel, Member member, boolean joined) {
@@ -153,7 +161,8 @@ public class DroneManager {
             if(text == null) return;
 
             if(joined) {
-                text.upsertPermissionOverride(member).grant(Permission.VIEW_CHANNEL).queue();
+                text.upsertPermissionOverride(member).grant(Permission.VIEW_CHANNEL)
+                        .and(channel.upsertPermissionOverride(member).grant(Permission.VOICE_CONNECT)).queue();
             } else if(voiceDrone.getOwnerId() != member.getIdLong()) {
                 text.getManager().removePermissionOverride(member).queue();
             }
@@ -169,10 +178,227 @@ public class DroneManager {
 
             if(voiceDrone != null) {
                 VoiceChannel voiceChannel = channel.getGuild().getVoiceChannelById(voiceDrone.getChannelId());
-                if(voiceChannel != null) voiceManager.createControlPanel(voiceChannel, true);
+                if(voiceChannel != null) this.createControlPanel(voiceChannel, true);
             }
         } catch (KeyNotFoundException e) {
             // Ignoring...
         }
+    }
+
+    public void checkToDeleteTemporary(VoiceChannel channel, boolean wasDeleted) {
+        long snowflakeId = channel.getIdLong();
+
+        try {
+            VoiceDrone drone = voiceDroneService.find(snowflakeId);
+
+            // Rule 1: If permanent, the drone will not be deleted
+            if (drone.isPermanent() && !wasDeleted) return;
+
+            int online = channel.getMembers().size();
+            if(online == 0) {
+                voiceDroneService.destroy(snowflakeId);
+
+                TextChannel textChannel = channel.getGuild().getTextChannelById(drone.getChatId());
+                if(textChannel != null) {
+                    textChannel.delete().and(channel.delete()).queue();
+                } else {
+                    channel.delete().queue();
+                }
+            }
+        } catch (KeyNotFoundException e) {
+            // Do nothing
+        }
+    }
+
+    public void checkToCreateTemporary(VoiceChannel channel, Member member) {
+
+        long snowflakeId = channel.getParentCategoryIdLong();
+
+        try {
+            // Checking if the current category is a dynamic category
+            VoiceHive hive = voiceHiveService.find(snowflakeId);
+
+            if(channel.getIdLong() != hive.getVoiceId()) return;
+
+            // Creating the voice drone
+            Category hiveCategory = channel.getParentCategory();
+            String droneName = this.getDroneName(member, hive);
+
+            // Create voice
+            VoiceChannel voice = hiveCategory.createVoiceChannel(droneName).complete();
+            TextChannel  text  = hiveCategory.createTextChannel(droneName).complete();
+
+            // Hiding channel
+            Role publicRole = channel.getGuild().getPublicRole();
+            Guild guild = voice.getGuild();
+            text.upsertPermissionOverride(publicRole).deny(Permission.VIEW_CHANNEL)
+                    .and(text.upsertPermissionOverride(member).grant(Permission.VIEW_CHANNEL))
+                    .and(voice.upsertPermissionOverride(member).grant(Permission.VOICE_CONNECT))
+                    .and(guild.moveVoiceMember(member, voice)).queue();
+
+            // Add voice to drone db
+            voiceDroneService.create(new VoiceDrone(voice.getIdLong(), text.getIdLong(), member.getIdLong(), null));
+            LOGGER.debug("Creating New Channel!");
+        } catch (KeyNotFoundException e) {
+            LOGGER.debug("Key not found, ignoring...");
+        }
+    }
+
+    private boolean isVisible(VoiceChannel channel) {
+        return channel.getGuild().getPublicRole().hasPermission(channel, Permission.VIEW_CHANNEL);
+    }
+
+    private boolean canConnect(VoiceChannel channel) {
+        return channel.getGuild().getPublicRole().hasPermission(channel, Permission.VOICE_CONNECT);
+    }
+
+    public void createControlPanel(VoiceChannel channel) throws KeyNotFoundException {
+        createControlPanel(channel, false);
+    }
+
+    public void createControlPanel(VoiceChannel channel, boolean forceSendInVoice) throws KeyNotFoundException {
+        VoiceDrone drone =  voiceDroneService.find(channel.getIdLong());
+        Member member    =  channel.getGuild().getMemberById(drone.getOwnerId());
+        GuildMessageChannel text = channel.getGuild().getTextChannelById(drone.getChatId());
+        text = text != null && !forceSendInVoice ? text : channel;
+        createControlPanel(channel, text, drone, member);
+    }
+
+    private void createControlPanel(VoiceChannel voiceChannel, GuildMessageChannel textChannel, VoiceDrone drone, Member member) {
+        LOGGER.debug("Creating control panel for: {}", voiceChannel.getId());
+
+        // region Embed Creator
+        EmbedBuilder builder = new EmbedBuilder();
+        builder.setTitle("⚙️ Control Panel - " + voiceChannel.getName());
+        builder.setDescription("Here, you can control your private voiceChannel.");
+        // Fields to Show
+        builder.addField("Owner", voiceChannel.getGuild().getMemberById(drone.getOwnerId()).getAsMention(), true);
+        builder.addField("User Limit", voiceChannel.getUserLimit() == 0 ? "No Limits" : String.valueOf(voiceChannel.getUserLimit()), true);
+        builder.addField("Visible", isVisible(voiceChannel) ? "Yes" : "No", true);
+        builder.addField("Connectable", canConnect(voiceChannel) ? "Yes" : "No", true);
+        builder.addField("Status", drone.isPermanent() ? "Permanent" : "Temporary", true);
+
+        // Tutorials
+        builder.addField("Invite User", "/dynamic invite @user to invite someone to your channel", false);
+        builder.addField("Kick User", "/dynamic kick @user to kick someone from your channel", false);
+        builder.addField("Ban User", "/dynamic ban @user to ban someone from your channel", false);
+
+        // Other things
+        builder.setColor(Color.YELLOW);
+        builder.setAuthor(member.getUser().getAsTag(), null, member.getUser().getAvatarUrl());
+
+        // endregion
+
+        // region Buttons
+        // General Config
+
+        net.dv8tion.jda.api.interactions.components.buttons.Button config     = net.dv8tion.jda.api.interactions.components.buttons.Button.primary(VoiceGroup.Dynamic.droneConfig           , "🔧 Settings");
+        net.dv8tion.jda.api.interactions.components.buttons.Button visibility = net.dv8tion.jda.api.interactions.components.buttons.Button.secondary(VoiceGroup.Dynamic.droneHideShow       , isVisible(voiceChannel) ? "👻 Hide" : "👀 Visible");
+        net.dv8tion.jda.api.interactions.components.buttons.Button connect    = net.dv8tion.jda.api.interactions.components.buttons.Button.secondary(VoiceGroup.Dynamic.dronePublicPrivate  , canConnect(voiceChannel) ? "📢 Public" : "🔒 Private");
+        net.dv8tion.jda.api.interactions.components.buttons.Button permanent  = Button.danger(VoiceGroup.Dynamic.dronePermTemp          , drone.isPermanent()     ? "⏳ Temporary" : "✨ Permanent");
+
+        ActionRow general = ActionRow.of(config, visibility, connect, permanent);
+        // endregion
+
+        // Send the message
+        sendControlPanel(textChannel, drone, builder, java.util.List.of(general));
+    }
+
+    private void sendControlPanel(GuildMessageChannel channel, VoiceDrone drone, EmbedBuilder builder, java.util.List<ActionRow> actionRows) {
+        // We already have a message, so we need to update it
+        if(drone.getControlPanel() != null) {
+            channel.editMessageEmbedsById(drone.getControlPanel(), builder.build()).setActionRows(actionRows).queue(q -> { /* It's ok! */}, e -> {
+                LOGGER.error("Error updating control panel: {}", e.getMessage());
+                // If error, we need to create a new one
+                sendNewControlPanel(channel, drone, builder, actionRows);
+            });
+        } else {
+            sendNewControlPanel(channel, drone, builder, actionRows);
+        }
+    }
+
+    private void sendNewControlPanel(GuildMessageChannel channel, VoiceDrone drone, EmbedBuilder builder, List<ActionRow> actionRows) {
+        channel.sendMessageEmbeds(builder.build()).setActionRows(actionRows).queue(q -> {
+            drone.setControlPanel(q.getIdLong());
+            try {
+                voiceDroneService.update(drone);
+            } catch (KeyNotFoundException ex) {
+                // WTF????
+                throw new RuntimeException(ex);
+            }
+        }, ee -> LOGGER.error("Error creating control panel: {}", ee.getMessage()));
+    }
+
+    public void makePermanent(VoiceChannel channel, boolean permanent) throws KeyNotFoundException {
+        VoiceDrone drone = voiceDroneService.find(channel.getIdLong());
+        drone.setPermanent(permanent);
+        voiceDroneService.update(drone);
+    }
+
+    private String getDroneName(Member member, VoiceHive hive) {
+        String droneName = VoiceManager.configModal_idle;
+        String username = member.getNickname() == null ? member.getEffectiveName() : member.getNickname();
+
+        Optional<Activity> streaming = member.getActivities().stream().filter(activity -> activity.getType() == Activity.ActivityType.STREAMING).findFirst();
+        Optional<Activity> playing   = member.getActivities().stream().filter(activity -> activity.getType() == Activity.ActivityType.PLAYING).findFirst();
+
+        if(streaming.isPresent() && !hive.getStreaming().isBlank()) {
+            droneName = hive.getStreaming().replaceAll("%CHANNEL%", streaming.get().getName());
+        }
+        if(playing.isPresent() && !hive.getPlaying().isBlank()) {
+            if(streaming.isEmpty()) droneName = hive.getPlaying().replaceAll("%PLAYING%", playing.get().getName());
+            else                    droneName = droneName.replaceAll("%PLAYING%", streaming.get().getName());
+        }
+        if(playing.isEmpty() && streaming.isEmpty()) {
+            droneName = hive.getIdle();
+        }
+
+        droneName = droneName.replaceAll("%OWNER%", username);
+
+        return droneName;
+    }
+
+    public void toggleDroneVisibility(Guild guild, MessageChannelUnion channel, Member member) throws MissingPermissionsException, KeyNotFoundException {
+        VoiceChannel voiceChannel = getVoiceChannel(guild, channel, member);
+
+        Role publicRole = guild.getPublicRole();
+
+        if(!isVisible(voiceChannel)) voiceChannel.upsertPermissionOverride(publicRole).grant(Permission.VIEW_CHANNEL).queue();
+        else                          voiceChannel.upsertPermissionOverride(publicRole).deny(Permission.VIEW_CHANNEL).queue();
+    }
+
+    public void toggleDronePublicPrivate(Guild guild, MessageChannelUnion channel, Member member) throws MissingPermissionsException, KeyNotFoundException {
+        VoiceChannel voiceChannel = getVoiceChannel(guild, channel, member);
+
+        Role publicRole = guild.getPublicRole();
+        if(!canConnect(voiceChannel)) voiceChannel.upsertPermissionOverride(publicRole).grant(Permission.VOICE_CONNECT).queue();
+        else                          voiceChannel.upsertPermissionOverride(publicRole).deny(Permission.VOICE_CONNECT).queue();
+    }
+
+    @Nullable
+    private VoiceChannel getVoiceChannel(Guild guild, MessageChannelUnion channel, Member member) throws KeyNotFoundException, MissingPermissionsException {
+        VoiceDrone drone;
+        if(channel.getType() == ChannelType.VOICE) drone = voiceDroneService.find(channel.getIdLong());
+        else                                       drone = voiceDroneService.findByChatId(channel.getIdLong());
+
+        if(!drone.getOwnerId().equals(member.getIdLong()) && !member.hasPermission(Permission.MANAGE_CHANNEL)) {
+            throw new MissingPermissionsException();
+        }
+
+        VoiceChannel voiceChannel = channel.getType() == ChannelType.VOICE ? channel.asVoiceChannel() : guild.getVoiceChannelById(drone.getChannelId());
+        return voiceChannel;
+    }
+
+    public void toggleDronePermTemp(Guild guild, MessageChannelUnion channel, Member member) throws MissingPermissionsException, KeyNotFoundException {
+        VoiceDrone drone;
+        if(channel.getType() == ChannelType.VOICE) drone = voiceDroneService.find(channel.getIdLong());
+        else                                       drone = voiceDroneService.findByChatId(channel.getIdLong());
+
+        if(!member.hasPermission(Permission.MANAGE_CHANNEL)) {
+            throw new MissingPermissionsException();
+        }
+        drone.setPermanent(!drone.isPermanent());
+        voiceDroneService.update(drone);
+        createControlPanel(guild.getVoiceChannelById(drone.getChannelId()));
     }
 }
